memory_management/pyproject.toml
[project]
name = "memory-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mem0ai>=0.1.116",
    "openai-agents>=0.2.10",
    "python-dotenv>=1.1.1",
]
